Example/IDA: Move `translatorForm` to `utils`
